radial chart files restructured
