+++ src/lab/Graphs/PassFailBar/base.ts
@@ -1,7 +1,5 @@
 export interface RadialChartMetric {
   value: number;
-  label: string;
-  baseColor?: string;
 }
 
 export interface RadialChartProps {
