+++ src/lab/Graphs/PassFailBar/testData.ts
@@ -2,8 +2,6 @@ import { RadialChartMetric } from "./base";
 
 const testRadialChartData: RadialChartMetric = {
   value: 60,
-  label: "pass",
-  baseColor: "#5B44BA",
 };
 
 export { testRadialChartData };
