--- src/lab/Graphs/RadialChart/index.ts
@@ -1,2 +1,0 @@
-export type { RadialChartMetric } from "./base";
-export * from "./RadialChart";
