+++ src/lab/Graphs/RadialGraphs/RadialChart/index.ts
@@ -1,0 +1,1 @@
+export * from "./RadialChart";
