+++ src/lab/Graphs/RadialGraphs/RadialProgressChart/RadialProgressChart.tsx
@@ -1,38 +1,32 @@
 import { useTheme } from "@material-ui/core";
 import { Arc, Group, ParentSize } from "@visx/visx";
 import React, { useState } from "react";
-import { LegendData, LegendTable } from "../LegendTable";
-import {
-  RadialChartChildProps,
-  RadialChartMetric,
-  RadialChartProps,
-} from "./base";
+import { LegendData } from "../../LegendTable";
+import { RadialChartMetric } from "../base";
+import { RadialChartChildProps, RadialChartProps } from "./base";
 import { useStyles } from "./styles";
 
-const RadialChartChild = ({
+const RadialProgressChartChild = ({
   width,
   height,
   radialData,
   arcWidth = 20,
   semiCircle = false,
-  legendTableHeight = 150,
-  showLegend = true,
   heading,
   circleExpandOnHover = 5,
-  alignLegendTable = "bottom",
-  showCenterHeading = true,
+  unit,
+  imageSrc,
+  imageAlt = "i",
   className,
 }: RadialChartChildProps) => {
   const { palette } = useTheme();
 
   let legenddata: Array<LegendData> = [{ data: [] }];
-  const [centerValue, setcenterValue] = useState<string>("0");
-  const [centerText, setCenterText] = useState<string>(heading ?? "");
+  let centerValue = "0";
+  const centerText = heading ?? "";
   const [currentHovered, setcurrentHovered] = useState<string>("");
-  const radialFigurWidth = alignLegendTable === "bottom" ? width : width / 2;
+  const radialFigurWidth = width;
   const circleOrient = semiCircle ? 1 : 2;
-
-  const scalerArc: number = circleOrient * Math.PI;
   const startAngle: number = -(Math.PI / 2);
   let currentAngle: number = startAngle;
   const outerRadius =
@@ -46,41 +40,34 @@ const RadialChartChild = ({
     width,
     height,
     circleOrient,
-    alignLegendTable,
-    legendTableHeight,
     innerRadius,
   });
-  const total = radialData
-    ? radialData.reduce(
-        (previousValue, currentValue) => previousValue + currentValue.value,
-        0
-      )
-    : NaN;
+  const total: number = radialData.value ? 100 : NaN;
+  const scalerArc: number = circleOrient * Math.PI;
+
   const radialArc: RadialChartMetric[] = radialData
-    ? radialData.map((elem) => {
-        return {
-          value: (total ? elem.value / total : 0) * scalerArc,
-          label: elem.label,
-          baseColor: elem.baseColor,
-        };
-      })
+    ? [
+        {
+          value: (total ? radialData.value / total : 0) * scalerArc,
+          label: radialData.label,
+          baseColor: radialData.baseColor,
+        },
+        {
+          value: (total ? (total - radialData.value) / total : 0) * scalerArc,
+          label: "rest",
+          baseColor: palette.disabledBackground,
+        },
+      ]
     : [{ value: NaN, label: "" }];
+
   if (centerValue === "0" && total > 0) {
-    setcenterValue(total.toString());
-    showCenterHeading ? setCenterText(heading ?? "") : null;
+    {
+      centerValue = radialData.value.toString();
+    }
   }
 
   legenddata = legenddata.splice(0);
 
-  if (radialData) {
-    radialData.map((element, index) => {
-      if (element.value !== undefined)
-        legenddata[index] = {
-          data: [element.label, element.value.toString()],
-          baseColor: element.baseColor,
-        };
-    });
-  }
   return width < 10 ? null : (
     <div className={`${classes.radialChartRoot} ${className}`}>
       <div className={classes.figureWithLegendTable}>
@@ -92,13 +79,7 @@ const RadialChartChild = ({
           />
 
           <Group
-            top={
-              circleOrient === 1
-                ? alignLegendTable === "right"
-                  ? height / 2 + innerRadius
-                  : height
-                : height / 2
-            }
+            top={circleOrient === 1 ? height : height / 2}
             left={radialFigurWidth / 2}
           >
             {total > 0 &&
@@ -122,16 +103,11 @@ const RadialChartChild = ({
                     startAngle={currentAngle}
                     endAngle={(currentAngle += elem.value)}
                     onMouseEnter={(e) => {
-                      setcenterValue(radialData[i].value.toString());
-                      showCenterHeading ? setCenterText(`${elem.label}`) : null;
                       setcurrentHovered(
                         e.currentTarget.getAttribute("id")?.toString() ?? ""
                       );
                     }}
                     onMouseLeave={() => {
-                      setcenterValue(total.toString());
-                      showCenterHeading ? setCenterText(`${heading}`) : null;
-
                       setcurrentHovered("");
                     }}
                     opacity={
@@ -160,40 +136,33 @@ const RadialChartChild = ({
             )}
           </Group>
         </svg>
-        {showLegend && (
-          <div className={classes.legendTableArea}>
-            <div className={classes.legendTableChild}>
-              <LegendTable data={legenddata} />
-            </div>
-          </div>
-        )}
       </div>
-
+      <div className={classes.centerIcon}>
+        <img src={imageSrc} alt={imageAlt} />
+      </div>
       <div className={classes.centerDataContainer}>
         <p className={`${classes.centerValue} ${classes.centerDataFont}`}>
-          {centerValue}
+          {centerValue + " " + unit}
         </p>
 
-        {showCenterHeading && (
-          <p className={`${classes.centerText} ${classes.centerDataFont}`}>
-            {centerText}
-          </p>
-        )}
+        <p className={`${classes.centerText} ${classes.centerDataFont}`}>
+          {centerText}
+        </p>
       </div>
     </div>
   );
 };
-const RadialChart: React.FC<RadialChartProps> = ({ ...rest }) => {
+const RadialProgressChart: React.FC<RadialChartProps> = ({ ...rest }) => {
   return (
     <ParentSize>
       {({ width, height }) =>
         width > 0 &&
         height > 0 && (
-          <RadialChartChild width={width} height={height} {...rest} />
+          <RadialProgressChartChild width={width} height={height} {...rest} />
         )
       }
     </ParentSize>
   );
 };
 
-export { RadialChart };
+export { RadialProgressChart };
