+++ src/lab/Graphs/RadialGraphs/RadialProgressChart/index.ts
@@ -1,0 +1,1 @@
+export * from "./RadialProgressChart";
