+++ src/lab/Graphs/RadialGraphs/RadialProgressChart/styles.ts
