+++ src/lab/Graphs/RadialGraphs/base.ts
@@ -1,0 +1,5 @@
+export interface RadialChartMetric {
+  value: number;
+  label: string;
+  baseColor?: string;
+}
