+++ src/lab/Graphs/RadialGraphs/index.ts
@@ -1,0 +1,3 @@
+export { RadialChartMetric } from "./base";
+export * from "./RadialChart";
+export * from "./RadialProgressChart";
