--- src/lab/Graphs/RadialProgressChart/index.ts
@@ -1,2 +1,0 @@
-export type { RadialChartMetric } from "./base";
-export * from "./RadialProgressChart";
