+++ src/lab/Graphs/index.ts
@@ -1,3 +1,3 @@
 export * from "./LegendTable";
 export * from "./LineAreaGraph";
-export * from "./RadialChart";
+export * from "./RadialGraphs";
